Add esm build to package
