+++ .gitignore
@@ -1,4 +1,6 @@
 .DS_Store
 node_modules
-out
+dist
+dist-cjs
+build
 coverage
